cpu_ctx not 32 bit dependant

+++ include/kernel/cpu_ctx.h
@@ -4,8 +4,8 @@
 struct cpu_ctx
 {
     // we just need the sp/ss, all other cpu context is stored on the stack.
-    uint32_t u_sp;
-    uint32_t k_sp;
+    unsigned long u_sp;
+    unsigned long k_sp;
 };
 
 #endif //CPU_CTX_H
